Move scenario fixture up into conftest
Make it accessible to all test mods.
Also tweak tests to account for new instance variable names.
